etl/requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
python-dotenv==1.0.1
tqdm==4.66.5